Moving language model data classes
