+++ src/python/PythonSDK/foundationallm/langchain/language_models/language_model_factory.py
@@ -1,7 +1,6 @@
 from langchain.base_language import BaseLanguageModel
 from foundationallm.config import Configuration
-from foundationallm.langchain.language_models import LanguageModelProviders, LanguageModelTypes
-
+from foundationallm.langchain.models.language_models import LanguageModelProviders, LanguageModelTypes
 from foundationallm.langchain.language_models.openai import AzureChatModel
 from foundationallm.langchain.language_models.openai import AzureTextCompletionModel
 from foundationallm.langchain.language_models.openai import OpenAIChatModel
