+++ src/python/PythonSDK/foundationallm/models/metadata/language_model.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, confloat
 from typing import Annotated
 
-from foundationallm.langchain.language_models import LanguageModelTypes, LanguageModelProviders
+from foundationallm.langchain.models.language_models import LanguageModelTypes, LanguageModelProviders
 
 class LanguageModel(BaseModel):
     """Language model metadata model."""
